% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimCoxIntervalCens.R
\name{AuxF}
\alias{AuxF}
\title{FUNCTION_TITLE}
\usage{
AuxF(tt, q1, q2, beta.q1 = log(2), beta.q2 = log(0.5), u)
}
\arguments{
\item{tt}{PARAM_DESCRIPTION}

\item{q1}{PARAM_DESCRIPTION}

\item{q2}{PARAM_DESCRIPTION}

\item{beta.q1}{PARAM_DESCRIPTION, Default: log(2)}

\item{beta.q2}{PARAM_DESCRIPTION, Default: log(0.5)}

\item{u}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
