% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcAuxAtPoint.R
\name{CalcAuxAtPoint}
\alias{CalcAuxAtPoint}
\title{FUNCTION_TITLE}
\usage{
CalcAuxAtPoint(w, w.res, point)
}
\arguments{
\item{w}{PARAM_DESCRIPTION}

\item{w.res}{PARAM_DESCRIPTION}

\item{point}{Scalar. The time of the risk set in the main analysis. In terms of the paper, t.}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
