% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcCoxCalibP.R
\name{CalcCoxCalibP}
\alias{CalcCoxCalibP}
\title{FUNCTION_TITLE}
\usage{
CalcCoxCalibP(w, w.res, point, fit.cox, hz.times, Q)
}
\arguments{
\item{w}{A matrix of time points when measurements on the binary covariate were obtained.}

\item{w.res}{A matrix of measurement results of the binary covariate. Each measurement corresponds to the time points in \code{w}}

\item{point}{PARAM_DESCRIPTION}

\item{fit.cox}{The result of \code{icenReg::ic_sp} on the interval-censored data}

\item{hz.times}{Times used for calculating the baseline hazard function from PH calibartion model}

\item{Q}{Matrix of covariates for PH calibration model}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
