% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcCoxCalibPderiv.R
\name{CalcCoxCalibPderiv}
\alias{CalcCoxCalibPderiv}
\title{FUNCTION_TITLE}
\usage{
CalcCoxCalibPderiv(w, w.res, point, fit.cox, hz.times, Q)
}
\arguments{
\item{w}{PARAM_DESCRIPTION}

\item{w.res}{PARAM_DESCRIPTION}

\item{point}{PARAM_DESCRIPTION}

\item{fit.cox}{PARAM_DESCRIPTION}

\item{hz.times}{PARAM_DESCRIPTION}

\item{Q}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[ICsurv]{Ispline}}
}
