% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcCoxCalibRSIntsP.R
\name{CalcCoxCalibRSIntsP}
\alias{CalcCoxCalibRSIntsP}
\title{FUNCTION_TITLE}
\usage{
CalcCoxCalibRSIntsP(w, w.res, point, fit.cox.rs.ints, hz.times, Q, pts.for.ints)
}
\arguments{
\item{w}{PARAM_DESCRIPTION}

\item{w.res}{PARAM_DESCRIPTION}

\item{point}{PARAM_DESCRIPTION}

\item{fit.cox.rs.ints}{PARAM_DESCRIPTION}

\item{hz.times}{PARAM_DESCRIPTION}

\item{Q}{PARAM_DESCRIPTION}

\item{pts.for.ints}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
