% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcGradEtaPers.R
\name{CalcGradEtaPers}
\alias{CalcGradEtaPers}
\title{FUNCTION_TITLE}
\usage{
CalcGradEtaPers(d1, d2, d3, Li, Ri, knots, order, eta.g, eta.b, Q)
}
\arguments{
\item{d1}{PARAM_DESCRIPTION}

\item{d2}{PARAM_DESCRIPTION}

\item{d3}{PARAM_DESCRIPTION}

\item{Li}{PARAM_DESCRIPTION}

\item{Ri}{PARAM_DESCRIPTION}

\item{knots}{PARAM_DESCRIPTION}

\item{order}{PARAM_DESCRIPTION}

\item{eta.g}{PARAM_DESCRIPTION}

\item{eta.b}{PARAM_DESCRIPTION}

\item{Q}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[ICsurv]{Ispline}}
}
