% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcGradEtaPersRSInts.R
\name{CalcGradEtaPersRSInts}
\alias{CalcGradEtaPersRSInts}
\title{FUNCTION_TITLE}
\usage{
CalcGradEtaPersRSInts(d1, d2, d3, Li, Ri, Q, fit.cox.rs.ints, pts.for.ints, tm,
  n.etas.per.fit)
}
\arguments{
\item{d1}{PARAM_DESCRIPTION}

\item{d2}{PARAM_DESCRIPTION}

\item{d3}{PARAM_DESCRIPTION}

\item{Li}{PARAM_DESCRIPTION}

\item{Ri}{PARAM_DESCRIPTION}

\item{Q}{PARAM_DESCRIPTION}

\item{fit.cox.rs.ints}{PARAM_DESCRIPTION}

\item{pts.for.ints}{PARAM_DESCRIPTION}

\item{tm}{PARAM_DESCRIPTION}

\item{n.etas.per.fit}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[ICsurv]{Ispline}}
}
