% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcNpmleRSP.R
\name{CalcNpmleRSP}
\alias{CalcNpmleRSP}
\title{FUNCTION_TITLE}
\usage{
CalcNpmleRSP(w, w.res, point, obs.tm)
}
\arguments{
\item{w}{A matrix of time points when measurements on the binary covariate were obtained.}

\item{w.res}{A matrix of measurement results of the binary covariate. Each measurement corresponds to the time points in \code{w}}

\item{point}{PARAM_DESCRIPTION}

\item{obs.tm}{Vector of observed main event time or censoring time}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[icenReg]{ic_np}}
}
