% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcSurvFromCox.R
\name{CalcSurvFromCox}
\alias{CalcSurvFromCox}
\title{FUNCTION_TITLE}
\usage{
CalcSurvFromCox(fit.cox, Qb, points, hz.times)
}
\arguments{
\item{fit.cox}{PARAM_DESCRIPTION}

\item{Qb}{PARAM_DESCRIPTION}

\item{points}{PARAM_DESCRIPTION}

\item{hz.times}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
