% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcVarTheta.R
\name{CalcVarNpmleRS}
\alias{CalcVarNpmleRS}
\title{Variance estimation and confidence intervals for the effect of the exposure under nonparametric  risk-set calibration models.}
\usage{
CalcVarNpmleRS(tm, event, w, w.res, BS = 100, CI = T)
}
\arguments{
\item{tm}{Vector of observed main event time or censoring time}

\item{event}{Vector of censoring indicators. \code{1} for event \code{0} for censored}

\item{w}{A matrix of time points when measurements on the binary covariate were obtained.}

\item{w.res}{A matrix of measurement results of the binary covariate. Each measurement corresponds to the time points in \code{w}}

\item{BS}{Number of bootstrap iterations, Default: 100}

\item{CI}{Should the function return confidencei intervals?, Default: T}
}
\value{
Variance estimate and possiblty confidence interval for the log hazard-ratio of the binary exposure under 
nonparametric risk-set calibration models.
}
\description{
Bootstrap calculations of the variance and confidence interval for the for the log hazard-ratio of the binary exposure
under nonparametric risk-set calibration models.
}
