% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcVarTheta.R
\name{CalcVarThetaWeib}
\alias{CalcVarThetaWeib}
\title{Variance estimation and confidence intervals for the effect of the exposure under Weibull calibration.}
\usage{
CalcVarThetaWeib(beta, etas, tm, event, ps, ps.deriv.shape, ps.deriv.scale, w,
  w.res)
}
\arguments{
\item{beta}{Coefficient of the binary covariate}

\item{etas}{shape and scale parameters of the Weibull calibration model.}

\item{tm}{Vector of observed main event time or censoring time}

\item{event}{Vector of censoring indicators. \code{1} for event \code{0} for censored}

\item{ps}{A matrix. Rows are observations, columns are time points of the events. The entry at the i-th row and j-column is
the conditional probability of positive exposure status for observation i at the j-th event time.}

\item{ps.deriv.shape}{The derivative of \code{ps} with respect to the shape parameter of the Weibull calibration model.}

\item{ps.deriv.scale}{The derivative of \code{ps} with respect to the scale parameter of the Weibull calibration model.}

\item{w}{A matrix of time points when measurements on the binary covariate were obtained.}

\item{w.res}{A matrix of measurement results of the binary covariate. Each measurement corresponds to the time points in \code{w}}
}
\value{
Variance estimate and possibly confidence interval for the log hazard-ratio of the binary exposure under
}
\description{
Bootstrap calculations of the variance and confidence interval for the for the log hazard-ratio of the binary exposure
under a Weibull calibration model.
}
