% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcWeibullCalibPderiv.R
\name{CalcWeibullCalibPderivScaleRS}
\alias{CalcWeibullCalibPderivScaleRS}
\title{FUNCTION_TITLE}
\usage{
CalcWeibullCalibPderivScaleRS(w, w.res, obs.tm, event, weib.rs.params)
}
\arguments{
\item{w}{PARAM_DESCRIPTION}

\item{w.res}{PARAM_DESCRIPTION}

\item{obs.tm}{PARAM_DESCRIPTION}

\item{event}{PARAM_DESCRIPTION}

\item{weib.rs.params}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[stats]{pweibull}}
}
