% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcWeibullRSP.R
\name{CalcWeibullRSP}
\alias{CalcWeibullRSP}
\title{Calculating the probabilities of positive binary exposure status at a given time point using risk-set Weibull calibration models}
\usage{
CalcWeibullRSP(w, w.res, point, weib.params)
}
\arguments{
\item{w}{A matrix of time points when measurements on the binary covariate were obtained.}

\item{w.res}{A matrix of measurement results of the binary covariate. Each measurement corresponds to the time points in \code{w}}

\item{point}{The time point at which the probabilities are estimated.}

\item{weib.params}{A bivariate vector. Shape and scale paramters of the Weibull calibration model.}
}
\value{
A vector of estimated probabilities of positive exposure status at time \code{point}.
}
\description{
For a given time point, calculate the probability of positive exposure value  for multiple observations (participants). 
The function uses the results of a Weibull calibration model fit, and given collected data on the history 
of the binary exposure for each participant.
}
\details{
At its present formm this function is identical to \code{CalcWeibullCalibP}. This is because the current version of the \code{ICcalib} package
(Version 1.0.005), the user loop over the main event times. Then, at each event time point, the user should include the appropaite Weibull
paramters as given by \code{FitCalibWeibullRS}.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[stats]{Weibull}}
}
