% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NaiveMethods.R
\name{CarryBack}
\alias{CarryBack}
\title{FUNCTION_TITLE}
\usage{
CarryBack(w, w.res, point)
}
\arguments{
\item{w}{PARAM_DESCRIPTION}

\item{w.res}{PARAM_DESCRIPTION}

\item{point}{the point in which a value for X is needed (i.e., current risk.set).}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
