% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OrdineryCalibration.R
\name{FitCalibCox}
\alias{FitCalibCox}
\title{FitCalibCox}
\usage{
FitCalibCox(w, w.res, Q, hz.times, n.int = 5, order = 2)
}
\arguments{
\item{w}{PARAM_DESCRIPTION}

\item{w.res}{PARAM_DESCRIPTION}

\item{Q}{PARAM_DESCRIPTION}

\item{hz.times}{PARAM_DESCRIPTION}

\item{n.int}{PARAM_DESCRIPTION, Default: 5}

\item{order}{PARAM_DESCRIPTION, Default: 2}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[ICsurv]{fast.PH.ICsurv.EM}}
}
