% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OrdineryCalibration.R
\name{FitCalibWeibull}
\alias{FitCalibWeibull}
\title{FitCalibWeibull}
\usage{
FitCalibWeibull(w, w.res)
}
\arguments{
\item{w}{PARAM_DESCRIPTION}

\item{w.res}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[fitdistrplus]{fitdistcens}}
 \code{\link[fitdistrplus]{fitdistcens}}
}
