% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitCalibWeibullRS.R
\name{FitCalibWeibullRS}
\alias{FitCalibWeibullRS}
\title{FUNCTION_TITLE}
\usage{
FitCalibWeibullRS(w, w.res, tm, event, lower = 1e-04, upper = 200)
}
\arguments{
\item{w}{A matrix of time points when measurements on the binary covariate were obtained.}

\item{w.res}{A matrix of measurement results of the binary covariate. Each measurement corresponds to the time points in \code{w}}

\item{tm}{PARAM_DESCRIPTION}

\item{event}{Vector of censoring indicators. \code{1} for event \code{0} for censored}

\item{lower}{PARAM_DESCRIPTION, Default: 1e-04}

\item{upper}{PARAM_DESCRIPTION, Default: 200}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[fitdistrplus]{fitdistcens}}
}
