% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeibAuxFunctions.R
\name{ICweibGrad}
\alias{ICweibGrad}
\title{FUNCTION_TITLE}
\usage{
ICweibGrad(etas, w, w.res)
}
\arguments{
\item{etas}{PARAM_DESCRIPTION}

\item{w}{PARAM_DESCRIPTION}

\item{w.res}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
