% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeibAuxFunctions.R
\name{ICweibGradRS}
\alias{ICweibGradRS}
\title{FUNCTION_TITLE}
\usage{
ICweibGradRS(etas.matrix, w, w.res, obs.tm, event)
}
\arguments{
\item{etas.matrix}{PARAM_DESCRIPTION}

\item{w}{PARAM_DESCRIPTION}

\item{w.res}{PARAM_DESCRIPTION}

\item{obs.tm}{PARAM_DESCRIPTION}

\item{event}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
