% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeibAuxFunctions.R
\name{ICweibLikLR}
\alias{ICweibLikLR}
\title{FUNCTION_TITLE}
\usage{
ICweibLikLR(weib.params, L, R)
}
\arguments{
\item{weib.params}{PARAM_DESCRIPTION}

\item{L}{PARAM_DESCRIPTION}

\item{R}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
