% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MidIdata.R
\name{MidIdata}
\alias{MidIdata}
\title{FUNCTION_TITLE}
\usage{
MidIdata(w, w.res, obs.tm, delta, Z = NULL)
}
\arguments{
\item{w}{PARAM_DESCRIPTION}

\item{w.res}{PARAM_DESCRIPTION}

\item{obs.tm}{PARAM_DESCRIPTION}

\item{delta}{PARAM_DESCRIPTION}

\item{Z}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
