% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OurICsurvPH.R
\name{OurICsurvPH}
\alias{OurICsurvPH}
\title{FUNCTION_TITLE}
\usage{
OurICsurvPH(d1, d2, d3, Li, Ri, Xp, n.int, order, g0, b0, tol, t.seq,
  equal = TRUE)
}
\arguments{
\item{d1}{PARAM_DESCRIPTION}

\item{d2}{PARAM_DESCRIPTION}

\item{d3}{PARAM_DESCRIPTION}

\item{Li}{PARAM_DESCRIPTION}

\item{Ri}{PARAM_DESCRIPTION}

\item{Xp}{PARAM_DESCRIPTION}

\item{n.int}{PARAM_DESCRIPTION}

\item{order}{PARAM_DESCRIPTION}

\item{g0}{PARAM_DESCRIPTION}

\item{b0}{PARAM_DESCRIPTION}

\item{tol}{PARAM_DESCRIPTION}

\item{t.seq}{PARAM_DESCRIPTION}

\item{equal}{PARAM_DESCRIPTION, Default: TRUE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[ICsurv]{Ispline}}
}
