% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OurICsurvPHgrad.R
\name{OurICsurvPHgrad}
\alias{OurICsurvPHgrad}
\title{FUNCTION_TITLE}
\usage{
OurICsurvPHgrad(Li, Ri, Xp, n.int, order, g1, b1, tol, t.seq, equal = TRUE)
}
\arguments{
\item{Li}{PARAM_DESCRIPTION}

\item{Ri}{PARAM_DESCRIPTION}

\item{Xp}{PARAM_DESCRIPTION}

\item{n.int}{PARAM_DESCRIPTION}

\item{order}{PARAM_DESCRIPTION}

\item{g1}{PARAM_DESCRIPTION}

\item{b1}{PARAM_DESCRIPTION}

\item{tol}{PARAM_DESCRIPTION}

\item{t.seq}{PARAM_DESCRIPTION}

\item{equal}{PARAM_DESCRIPTION, Default: TRUE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
