% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimCoxIntervalCens.R
\name{SimCoxIntervalCensCoxNonTerminal}
\alias{SimCoxIntervalCensCoxNonTerminal}
\title{FUNCTION_TITLE}
\usage{
SimCoxIntervalCensCoxNonTerminal(n.sample, lambda = 0.25, alpha = 0.1,
  beta0 = log(0.75), gamma.q = c(log(0.75), log(2.5)), gamma.z = log(1.5),
  mu = 1, n.points = 5)
}
\arguments{
\item{n.sample}{PARAM_DESCRIPTION}

\item{lambda}{PARAM_DESCRIPTION, Default: 0.25}

\item{alpha}{PARAM_DESCRIPTION, Default: 0.1}

\item{beta0}{PARAM_DESCRIPTION, Default: log(0.75)}

\item{gamma.q}{PARAM_DESCRIPTION, Default: c(log(0.75), log(2.5))}

\item{gamma.z}{PARAM_DESCRIPTION, Default: log(1.5)}

\item{mu}{PARAM_DESCRIPTION, Default: 1}

\item{n.points}{PARAM_DESCRIPTION, Default: 5}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
