% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimCoxIntervalCens.R
\name{SimCoxIntervalCensSingle}
\alias{SimCoxIntervalCensSingle}
\title{FUNCTION_TITLE}
\usage{
SimCoxIntervalCensSingle(n.sample, lambda = 0.25, alpha = NULL,
  beta0 = log(0.75), mu = 1, n.points = 5, weib.shape = 1,
  weib.scale = 1)
}
\arguments{
\item{n.sample}{PARAM_DESCRIPTION}

\item{lambda}{PARAM_DESCRIPTION, Default: 0.25}

\item{alpha}{PARAM_DESCRIPTION, Default: NULL}

\item{beta0}{PARAM_DESCRIPTION, Default: log(0.75)}

\item{mu}{PARAM_DESCRIPTION, Default: 1}

\item{n.points}{PARAM_DESCRIPTION, Default: 5}

\item{weib.shape}{PARAM_DESCRIPTION, Default: 1}

\item{weib.scale}{PARAM_DESCRIPTION, Default: 1}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
