% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcNpmleCalibP.R
\name{CalcNpmleCalibP}
\alias{CalcNpmleCalibP}
\title{FUNCTION_TITLE}
\usage{
CalcNpmleCalibP(w, w.res, point, fit.npmle)
}
\arguments{
\item{w}{A matrix of time points when measurements on the binary covariate were obtained.}

\item{w.res}{A matrix of measurement results of the binary covariate. Each measurement corresponds to the time points in \code{w}}

\item{point}{PARAM_DESCRIPTION}

\item{fit.npmle}{The result of \code{icenReg::ic_np} on the interval-censored data}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
