% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcNpmleRSP.R
\name{CalcNpmleRSP}
\alias{CalcNpmleRSP}
\title{Calculating the probabilities of positive binary exposure status at a given time point using a nonparametric risk-set calibration models}
\usage{
CalcNpmleRSP(w, w.res, point, obs.tm)
}
\arguments{
\item{w}{A matrix of time points when measurements on the binary covariate were obtained.}

\item{w.res}{A matrix of measurement results of the binary covariate. Each measurement corresponds to the time points in \code{w}}

\item{point}{The time point at which the probabilities are estimated}

\item{obs.tm}{Vector of observed main event time or censoring time}
}
\value{
A vector of estimated probabilities of positive exposure status at time \code{point}.
}
\description{
For a given time point, calculate the probability of positive exposure value  for multiple observations (participants). 
The function first fits the nonparametric riskset calibration models at each main event time point and then calculates the probabilities
of positive binary exposure status.
}
\details{
Unlike other risk-set calibration models, this function carries both the NPMLE at each main event time point and provides
the estimated probabilities for positive exposure status at time \code{point}.
}
\seealso{
\code{\link[icenReg]{ic_np}}
}
