% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcVarTheta.R
\name{CalcVarNpmle}
\alias{CalcVarNpmle}
\title{FUNCTION_TITLE}
\usage{
CalcVarNpmle(tm, event, w, w.res, BS = 100, CI = T)
}
\arguments{
\item{tm}{PARAM_DESCRIPTION}

\item{event}{Vector of censoring indicators. \code{1} for event \code{0} for censored}

\item{w}{A matrix of time points when measurements on the binary covariate were obtained.}

\item{w.res}{A matrix of measurement results of the binary covariate. Each measurement corresponds to the time points in \code{w}}

\item{BS}{Number of bootstrap iterations, Default: 100}

\item{CI}{PARAM_DESCRIPTION, Default: T}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[stats]{optimize}},\code{\link[stats]{var}},\code{\link[stats]{quantile}}
}
