% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcVarTheta.R
\name{CalcVarParam}
\alias{CalcVarParam}
\title{Variance estimation for main proportional hazards model, when the calibration model is also PH}
\usage{
CalcVarParam(theta, tm, event, Z, Q, ps, ps.deriv, w, w.res, fit.cox)
}
\arguments{
\item{theta}{Coefficient vector from main PH model. First coefficient corresponds to X, the rest to Z}

\item{tm}{Vector of observed main event time or censoring time}

\item{event}{Vector of censoring indicators. \code{1} for event \code{0} for censored}

\item{Z}{Additional variables for the main model other than the binary covariate}

\item{Q}{Matrix of covariate for PH calibration model}

\item{ps}{A matrix. Rows are observations, columns are time points of the events. The entry at the i-th row and j-column is
the conditional probability of positive exposure status for observation i at the j-th event time.}

\item{ps.deriv}{A matrix. Rows are observations, columns are time points of the events.  The derivative of \code{ps} with 
respect to the calibration model parameters}

\item{w}{A matrix of time points when measurements on the binary covariate were obtained.}

\item{w.res}{A matrix of measurement results of the binary covariate. Each measurement corresponds to the time points in \code{w}}

\item{fit.cox}{The result of \code{icenReg::ic_sp} on the interval-censored data}
}
\value{
The covariance matrix. The first row and column are for the binary exposure.
}
\description{
Calculates the covariance matrix for the parameters of the main calibration model. This includes the variance 
of the binary exposure estimate and the other covariates.
}
