% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcVarTheta.R
\name{CalcVarParamRSInts}
\alias{CalcVarParamRSInts}
\title{FUNCTION_TITLE}
\usage{
CalcVarParamRSInts(theta, tm, event, Z, Q, ps, ps.deriv, w, w.res,
  fit.cox.rs.ints, pts.for.ints, n.etas.per.fit)
}
\arguments{
\item{theta}{Coefficeint vector from main PH model. Fist coefficent corresponds to `X`, the rest to `Z`}

\item{tm}{PARAM_DESCRIPTION}

\item{event}{Vector of censoring indicators. \code{1} for event \code{0} for censored}

\item{Z}{Additional variables for the main model other than the binary covaraite}

\item{Q}{Matrix of covariates for PH calibration model}

\item{ps}{A matrix. Rows are observations, columns are time points of the events. $ps(i,j)=Pr(X_i(time of case j)=1|history)$}

\item{ps.deriv}{A matrix. Rows are observations, columns are time points of the events. $ps.deriv(i,j)$ is the derivative of Pr(X_i(time of case j)=1|history)}

\item{w}{A matrix of time points when measurements on the binary covariate were obtained.}

\item{w.res}{A matrix of measurement results of the binary covariate. Each measurement corresponds to the time points in \code{w}}

\item{fit.cox.rs.ints}{The result of \code{FitCalibCoxRSInts} on the interval-censored data}

\item{pts.for.ints}{Points defining the intervals for grouping risk-sets (first one has to be zero). Should be sorted from zero up}

\item{n.etas.per.fit}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[MASS]{ginv}}
}
