% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcVarTheta.R
\name{CalcVarThetaWeib}
\alias{CalcVarThetaWeib}
\title{FUNCTION_TITLE}
\usage{
CalcVarThetaWeib(beta, etas, tm, event, ps, ps.deriv.shape, ps.deriv.scale, w,
  w.res)
}
\arguments{
\item{beta}{Coefficient of the binary covariate}

\item{etas}{PARAM_DESCRIPTION}

\item{tm}{PARAM_DESCRIPTION}

\item{event}{Vector of censoring indicators. \code{1} for event \code{0} for censored}

\item{ps}{A matrix. Rows are observations, columns are time points of the events. $ps(i,j)=Pr(X_i(time of case j)=1|history)$}

\item{ps.deriv.shape}{Same as `ps.deriv.scale` but for the shape parameter}

\item{ps.deriv.scale}{A matrix. Rows are observations, columns are time points of the events. $ps.deriv.scale(i,j)$ is the derivative}

\item{w}{A matrix of time points when measurements on the binary covariate were obtained.}

\item{w.res}{A matrix of measurement results of the binary covariate. Each measurement corresponds to the time points in \code{w}}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[numDeriv]{hessian}}
}
