% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcVarTheta.R
\name{CalcVarThetaWeibRS}
\alias{CalcVarThetaWeibRS}
\title{FUNCTION_TITLE}
\usage{
CalcVarThetaWeibRS(beta, etas.matrix, tm, event, ps.rs, ps.deriv.shape.rs,
  ps.deriv.scale.rs, w, w.res)
}
\arguments{
\item{beta}{Coefficient of the binary covariate}

\item{etas.matrix}{PARAM_DESCRIPTION}

\item{tm}{PARAM_DESCRIPTION}

\item{event}{Vector of censoring indicators. \code{1} for event \code{0} for censored}

\item{ps.rs}{PARAM_DESCRIPTION}

\item{ps.deriv.shape.rs}{Same as `ps.deriv.scale` but for the risk-set shape parameters}

\item{ps.deriv.scale.rs}{A matrix. Rows are observations, columns are time points of the events. $ps.deriv(i,j)$ is the derivative}

\item{w}{A matrix of time points when measurements on the binary covariate were obtained.}

\item{w.res}{A matrix of measurement results of the binary covariate. Each measurement corresponds to the time points in \code{w}}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
