% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcWeibullRSP.R
\name{CalcWeibullRSP}
\alias{CalcWeibullRSP}
\title{FUNCTION_TITLE}
\usage{
CalcWeibullRSP(w, w.res, point, weib.params)
}
\arguments{
\item{w}{A matrix of time points when measurements on the binary covariate were obtained.}

\item{w.res}{A matrix of measurement results of the binary covariate. Each measurement corresponds to the time points in \code{w}}

\item{point}{PARAM_DESCRIPTION}

\item{weib.params}{A bivariate vector. Shape and scale paramters of the Weibull calibration model.}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[stats]{pweibull}}
}
