% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OrdinaryCalibration.R
\name{FitCalibCox}
\alias{FitCalibCox}
\title{Fitting Proportional Hazards Calibration Models with Covariates}
\usage{
FitCalibCox(w, w.res, Q, hz.times, n.int = 5, order = 2)
}
\arguments{
\item{w}{A matrix of time points when measurements on the binary covariate were obtained.}

\item{w.res}{A matrix of measurement results of the binary covariate. Each measurement corresponds to the time points in \code{w}}

\item{Q}{Matrix of covariates for PH calibration model}

\item{hz.times}{Times used for calculating the baseline hazard function from PH calibration model}

\item{n.int}{The number of interior knots to be used, see \code{ICsurv::fast.PH.ICsurv.EM}, Default: 5}

\item{order}{the order of the basis functions. See \code{ICsurv::fast.PH.ICsurv.EM}, Default: 2}
}
\value{
An object created by ICsurv::fast.PH.ICsurv.EM, with additional variables \code{knots} and \code{order}.
}
\description{
Fits a proportional hazards calibration model for time-to-exposure from interval-censored data with covariates. The exposure is a binary covariate measured
in intermittent times. The covariates (\code{Q}) are associated with the time-to-exposure.
}
\seealso{
\code{\link[ICsurv]{fast.PH.ICsurv.EM}}
}
