% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitCalibCoxRS.R
\name{FitCalibCoxRS}
\alias{FitCalibCoxRS}
\title{Fitting Proportional Hazards Risk-Set Calibration Models with Covariates}
\usage{
FitCalibCoxRS(w, w.res, Q, hz.times, obs.tm, n.int = 5, order = 2, event)
}
\arguments{
\item{w}{A matrix of time points when measurements on the binary covariate were obtained.}

\item{w.res}{A matrix of measurement results of the binary covariate. Each measurement corresponds to the time points in \code{w}}

\item{Q}{Matrix of covariates for PH calibration model}

\item{hz.times}{Times used for calculating the baseline hazard function from PH calibartion model}

\item{obs.tm}{Vector of observed main event time or censoring time}

\item{n.int}{The number of interior knots to be used, see \code{ICsurv::fast.PH.ICsurv.EM}, Default: 5}

\item{order}{the order of the basis functions. See \code{ICsurv::fast.PH.ICsurv.EM}, Default: 2}

\item{event}{Vector of censoring indicators. \code{1} for event \code{0} for censored}
}
\value{
A list of Cox PH model fits, each supplmented with the knots and order used for the I-splines.
}
\description{
Fits proportional hazards risk-set calibration models for time-to-exposure from interval-censored data with covariates. The exposure is a binary covariate measured
in intermittent times. The covariates (\code{Q}) are associated with the time-to-exposure. This function fits a calibration model at each main event time point,
using only members of the risk set at that time point.
}
\details{
In case of an error in the model-fitting at a certain time point, a proportional hazards calibration 
model is fitted (for all the data) and used for that time point.
}
\seealso{
\code{\link[ICsurv]{fast.PH.ICsurv.EM}},  \code{\link[ICcalib]{FitCalibCox}}
}
