% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitCalibCoxRSInts.R
\name{FitCalibCoxRSInts}
\alias{FitCalibCoxRSInts}
\title{Fitting Proportional Hazards Grouped Risk-Set Calibration Models with Covariates}
\usage{
FitCalibCoxRSInts(w, w.res, Q, hz.times, n.int = 5, order = 2, tm, event,
  pts.for.ints)
}
\arguments{
\item{w}{A matrix of time points when measurements on the binary covariate were obtained.}

\item{w.res}{A matrix of measurement results of the binary covariate. Each measurement corresponds to the time points in \code{w}}

\item{Q}{Matrix of covariates for PH calibration model}

\item{hz.times}{Times used for calculating the baseline hazard function of a PH calibration model}

\item{n.int}{The number of interior knots to be used, see \code{ICsurv::fast.PH.ICsurv.EM}, Default: 5}

\item{order}{the order of the basis functions. See \code{ICsurv::fast.PH.ICsurv.EM}, Default: 2}

\item{tm}{Vector of observed main event time or censoring time}

\item{event}{Vector of censoring indicators. \code{1} for event \code{0} for censored}

\item{pts.for.ints}{Points defining the intervals for grouping risk-sets (first one has to be zero). Should be sorted from zero up}
}
\value{
A list of Cox PH model fits, each supplemented with the knots and order used for the I-splines.
}
\description{
Fits proportional hazards grouped risk-set calibration models for time-to-exposure from interval-censored data with covariates. The exposure is a binary covariate measured
in intermittent times. The covariates (\code{Q}) are associated with the time-to-exposure. Unlike \code{FitCalibCoxRS}, this function fits a calibration model
at each of the given points for \code{pts.for.ints}.
}
\details{
In case of an error in the model-fitting at a certain time point, a proportional hazards calibration 
model (for all the data) is fitted and used for that time point.
}
\seealso{
\code{\link[ICsurv]{fast.PH.ICsurv.EM}},  \code{\link[ICcalib]{FitCalibCox}}
}
