% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OrdinaryCalibration.R
\name{FitCalibWeibull}
\alias{FitCalibWeibull}
\title{Fitting Weibull Calibration Models}
\usage{
FitCalibWeibull(w, w.res)
}
\arguments{
\item{w}{A matrix of time points when measurements on the binary covariate were obtained.}

\item{w.res}{A matrix of measurement results of the binary covariate. Each measurement corresponds to the time points in \code{w}}
}
\value{
A bivariate vector: the estimated Weibull shape and scale parameters
}
\description{
Fits a Weibull calibration model for time-to-exposure from interval-censored data. The exposure is a binary covariate measured
in intermittent times.
}
\seealso{
\code{\link[fitdistrplus]{fitdistcens}}
}
