% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitCalibWeibullRS.R
\name{FitCalibWeibullRS}
\alias{FitCalibWeibullRS}
\title{FitCalibWeibullRS}
\usage{
FitCalibWeibullRS(w, w.res, tm, event, lower = 1e-04, upper = 200)
}
\arguments{
\item{w}{A matrix of measurement times: each row is a participant, each entry is meausrment time, see example below}

\item{w.res}{A matrix of exposure measurments corresponding to the matrix w}

\item{tm}{PARAM_DESCRIPTION}

\item{event}{PARAM_DESCRIPTION}

\item{lower}{PARAM_DESCRIPTION, Default: 1e-04}

\item{uppper}{PARAM_DESCRIPTION, Default: 200}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Fit a parametric Weibull calibration model for time-to-exposure
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[fitdistrplus]{fitdistcens}}
 \code{\link[fitdistrplus]{fitdistcens}}
}
