% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MidIdata.R
\name{MidIdata}
\alias{MidIdata}
\title{FUNCTION_TITLE}
\usage{
MidIdata(w, w.res, obs.tm, delta, Z = NULL)
}
\arguments{
\item{w}{A matrix of time points when measurements on the binary covariate were obtained.}

\item{w.res}{A matrix of measurement results of the binary covariate. Each measurement corresponds to the time points in \code{w}}

\item{obs.tm}{Vector of observed main event time or censoring time}

\item{delta}{Vector of censoring indicators. \code{1} for event \code{0} for censored}

\item{Z}{Additional variables for the main model other than the binary covaraite, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
